# docs/openapi.json
{
  "components": {
    "schemas": {
      "ActorType": {
        "enum": [
          "PROVIDER_PORTAL",
          "VOICE_AGENT",
          "SYSTEM",
          "HUMAN_REVIEWER"
        ],
        "title": "ActorType",
        "type": "string"
      },
...
      "AuditEventView": {
        "properties": {
          "actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "actor_type": {
            "$ref": "#/components/schemas/ActorType"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "event_type": {
            "$ref": "#/components/schemas/AuditEventType"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sequence",
          "event_type",
          "actor_type",
          "actor_id",
          "occurred_at",
          "request_id",
          "data"
        ],
        "title": "AuditEventView",
        "type": "object"
      },
...
      "CloseCaseCommand": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reason": {
            "$ref": "#/components/schemas/CloseReason"
          }
        },
        "required": [
          "reason"
        ],
        "title": "CloseCaseCommand",
        "type": "object"
      },
      "CloseReason": {
        "enum": [
          "WITHDRAWN_BY_PROVIDER",
          "DECISION_COMMUNICATED"
        ],
        "title": "CloseReason",
        "type": "string"
      },
      "CreateCaseCommand": {
        "additionalProperties": false,
        "properties": {
          "information": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseInformationUpdate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "CreateCaseCommand",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nBackend for provider pre-authorisation requests: case intake, validation, rule evaluation, advisory\nrecommendations, human review, and an immutable audit trail.\n\n**The system never issues final authorisations or denials.** Only an authorised human reviewer can move a case\nto `APPROVED` or `DENIED`.\n\n**Authentication assumption:** requests arrive through a gateway that authenticates the caller and sets\n`X-Actor-Type`, `X-Actor-Id`, and (for reviewers) `X-Actor-Roles`. The service must not be exposed directly.\n\nEvery response carries `X-Request-ID`. Errors use a uniform envelope: `{\"error\": {\"code\", \"message\", \"details\",\n\"request_id\", \"case_id\"}}`.\n",
    "title": "Pre-Authorisation Case Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "post": {
        "description": "Opens a new case, optionally with initial information. Returns the case with a short, speakable `case_reference` suitable for quoting to a caller.\n\n**Authorisation:** Actor type `PROVIDER_PORTAL` or `VOICE_AGENT`.\n\n**State transitions:** \u2192 `RECEIVED`; if information is supplied, `RECEIVED` \u2192 `INFORMATION_COLLECTION`.",
        "operationId": "create_case_api_v1_cases_post",
        "parameters": [
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTAKE_ACTOR_REQUIRED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED / UNKNOWN_PROVIDER / MEMBER_NOT_VERIFIED / UNKNOWN_POLICY / POLICY_MEMBER_MISMATCH / UNKNOWN_PROCEDURE / SERVICE_DATE_IN_PAST"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Create a pre-authorisation case",
        "tags": [
          "Provider interaction"
        ]
      }
    },
    "/api/v1/cases/by-reference/{case_reference}": {
      "get": {
        "description": "Resolves a caller-quoted case reference (e.g. `PA-7K3M9Q2B`).\n\n**Authorisation:** Any authenticated actor.\n\n**State transitions:** None.",
        "operationId": "get_case_by_reference_api_v1_cases_by_reference__case_reference__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_reference",
            "required": true,
            "schema": {
              "pattern": "^PA-[0-9A-Z]{8}$",
              "title": "Case Reference",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Look up a case by its reference",
        "tags": [
          "Provider interaction"
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "description": "Returns all collected information, documents, and current status.\n\n**Authorisation:** Any authenticated actor.\n\n**State transitions:** None.",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Retrieve a case",
        "tags": [
          "Provider interaction"
        ]
      }
    },
    "/api/v1/cases/{case_id}/audit-events": {
      "get": {
        "description": "Immutable, ordered audit events for the case. Events are append-only at the database level.\n\n**Authorisation:** Actor type `HUMAN_REVIEWER` or `SYSTEM`.\n\n**State transitions:** None.",
        "operationId": "audit_events_api_v1_cases__case_id__audit_events_get",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventView"
                  },
                  "title": "Response Audit Events Api V1 Cases  Case Id  Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HUMAN_REVIEWER_REQUIRED"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Retrieve the audit trail",
        "tags": [
          "Audit"
        ]
      }
    },
    "/api/v1/cases/{case_id}/closure": {
      "post": {
        "description": "`WITHDRAWN_BY_PROVIDER` before review; `DECISION_COMMUNICATED` after a human decision.\n\n**Authorisation:** Withdrawal: `PROVIDER_PORTAL` or `VOICE_AGENT`. After a decision: `SYSTEM` or `HUMAN_REVIEWER`.\n\n**State transitions:** `RECEIVED` / `INFORMATION_COLLECTION` / `PENDING_INFORMATION` / `RECOMMENDATION_READY` / `APPROVED` / `DENIED` \u2192 `CLOSED` (terminal).",
        "operationId": "close_case_api_v1_cases__case_id__closure_post",
        "parameters": [
          {
            "description": "Case UUID",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case UUID",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Verified actor type set by the gateway",
            "in": "header",
            "name": "x-actor-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor type set by the gateway",
              "title": "X-Actor-Type"
            }
          },
          {
            "description": "Verified actor identifier set by the gateway",
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Verified actor identifier set by the gateway",
              "title": "X-Actor-Id"
            }
          },
          {
            "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
            "in": "header",
            "name": "x-actor-roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated reviewer roles (HUMAN_REVIEWER only)",
              "title": "X-Actor-Roles"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseCaseCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "ACTOR_REQUIRED / INVALID_ACTOR"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "TRANSITION_NOT_AUTHORIZED"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CASE_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_CLOSE_REASON / INVALID_STATE_TRANSITION / CONCURRENT_MODIFICATION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "REQUEST_VALIDATION_FAILED"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR / INTEGRITY_VIOLATION"
          }
        },
        "summary": "Close a case",
        "tags": [
          "Provider interaction"
        ]
      }
    },
...
  }
}